Move SystemClipboard to ui package
