+++ game-core/src/main/java/games/strategy/debug/console/window/ConsoleConfiguration.java
@@ -3,7 +3,7 @@ package games.strategy.debug.console.window;
 import java.util.logging.Handler;
 import java.util.logging.LogManager;
 
-import games.strategy.util.SystemClipboard;
+import games.strategy.ui.SystemClipboard;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
